+++ ui/components/AutoCompleteField/AutoCompleteField.js
@@ -5,12 +5,13 @@ import { debounce } from "lodash";
 import onInputValueChangeService from "./onInputValueChangeService";
 import highlightItem from "services/hightlightItem";
 import ReactHtmlParser from "react-html-parser";
+import { Spinner } from "reactstrap";
 
 let debouncedOnInputValueChange = null;
 
 // Permet de sélectionner un élément dans la liste d'items correspondant à un texte entré au clavier
 export const compareAutoCompleteValues = (items, value) => {
-  return items.findIndex((element) => (element.label ? element.label.toLowerCase() === value.toLowerCase() : false));
+  return items.findIndex((element) => (element?.label ? element.label.toLowerCase() === value.toLowerCase() : false));
 };
 
 // indique l'attribut de l'objet contenant le texte de l'item sélectionné à afficher
@@ -30,6 +31,7 @@ export const AutoCompleteField = ({
   scrollParentId,
   previouslySelectedItem,
   illustration,
+  searchPlaceholder,
   ...props
 }) => {
   useEffect(() => {
@@ -41,6 +43,7 @@ export const AutoCompleteField = ({
         items,
         setInputItems,
         selectItem,
+        setLoadingState,
         onInputValueChangeFunction,
         compareItemFunction,
         onSelectedItemChangeFunction,
@@ -54,6 +57,7 @@ export const AutoCompleteField = ({
 
   const [inputItems, setInputItems] = useState(items);
   const [initialized, setInitialized] = useState(false);
+  const [loadingState, setLoadingState] = useState("loading");
 
   const itemToString = (item) => {
     if (itemToStringFunction) return item ? itemToStringFunction(item) : "";
@@ -61,9 +65,8 @@ export const AutoCompleteField = ({
   };
 
   // hack pour scroller un champ autocomplete dont les valeurs pourraient être cachées par le clavier du mobile
-  const onFocus = (e) => {
+  const onFocusTriggered = (e) => {
     let ancestor = e.currentTarget.closest(`#${scrollParentId}`);
-
     if (ancestor) {
       setTimeout(() => {
         if (typeof window !== "undefined") {
@@ -81,11 +84,13 @@ export const AutoCompleteField = ({
     highlightedIndex,
     getItemProps,
     selectItem,
+    openMenu,
     inputValue,
   } = useCombobox({
-    id: 'lang-switcher',
+    id: "lang-switcher",
     items: inputItems,
     itemToString,
+    defaultHighlightedIndex: 0,
     initialSelectedItem: previouslySelectedItem,
     initialIsOpen,
     onSelectedItemChange: ({ selectedItem }) => {
@@ -95,6 +100,9 @@ export const AutoCompleteField = ({
       }
     },
     onInputValueChange: async ({ inputValue }) => {
+      if (loadingState === "done") {
+        setLoadingState("loading");
+      }
       if (!debouncedOnInputValueChange) {
         debouncedOnInputValueChange = debounce(onInputValueChangeService, 300);
       }
@@ -103,6 +111,7 @@ export const AutoCompleteField = ({
         inputItems,
         items,
         setInputItems,
+        setLoadingState,
         selectItem,
         onInputValueChangeFunction,
         compareItemFunction,
@@ -110,40 +119,66 @@ export const AutoCompleteField = ({
     },
   });
 
-  const classesOfContainer = props?.isHome ? '' : 'c-logobar-formgroup'
-  const classesOfInsider = props?.isHome ? 'form-control-lg w-100 c-input-work' : 'c-logobar-field'
+  const classesOfContainer = props?.isHome ? "" : "c-logobar-formgroup";
+  const classesOfInsider = props?.isHome ? "form-control-lg w-100 c-input-work" : "c-logobar-field";
 
   return (
     <div className="">
       <div className={`c-input-work-container ${classesOfContainer}`} {...getComboboxProps()}>
         <label className="c-logobar-label">{kind}</label>
         <input
-          {...getInputProps()}
+          {...getInputProps({
+            onFocus: (e) => {
+              if (!isOpen) {
+                openMenu();
+              }
+              onFocusTriggered(e);
+            },
+          })}
           className={`${classesOfInsider} ${
             inputValue && inputValue.length > 20 ? "is-text-too-long" : "is-text-not-too-long"
           }`}
           placeholder={props.placeholder}
-          onFocus={onFocus}
           name={props.name}
           aria-describedby="name"
         />
         {illustration && <img className="c-input-work-img" src={illustration} alt="" />}
       </div>
-      <ul {...getMenuProps()} className="c-autocomplete__menu">
-        {isOpen &&
-          inputItems.map((item, index) =>
-            item.label ? (
-              <li
-                className={highlightedIndex === index ? "c-autocomplete__option--highlighted" : ""}
-                key={`${index}`}
-                {...getItemProps({ item: item.label, index })}
-              >
-                {ReactHtmlParser(highlightItem(item.label, inputValue))}
-              </li>
-            ) : (
-              ""
-            )
-          )}
+
+      <ul {...getMenuProps()} className={`c-autocomplete__menu is-open-${isOpen}`}>
+        {(() => {
+          if (isOpen) {
+            if (inputValue.length === 0) {
+              return <li className="c-autocomplete-neutral">{searchPlaceholder}</li>;
+            } else if (loadingState === "loading") {
+              return (
+                <li className="c-autocomplete-neutral">
+                  <Spinner color="primary" />
+                  &nbsp;Veuillez patienter
+                </li>
+              );
+            } else if (inputValue.length > 0 && inputItems?.length === 0) {
+              return <li className="c-autocomplete-neutral">Pas de résultat, veuillez modifier votre recherche</li>;
+            } else {
+              return (
+                <>
+                  <li className="c-autocomplete-minititle">Résultats de votre recherche</li>
+                  {inputItems
+                    .filter((item) => !!item?.label)
+                    .map((item, index) => (
+                      <li
+                        className={highlightedIndex === index ? "c-autocomplete__option--highlighted" : ""}
+                        key={`${index}`}
+                        {...getItemProps({ item: item.label, index })}
+                      >
+                        {ReactHtmlParser(highlightItem(item.label, inputValue))}
+                      </li>
+                    ))}
+                </>
+              );
+            }
+          }
+        })()}
       </ul>
     </div>
   );