+++ ui/components/HeaderForm/HeaderForm.js
@@ -28,6 +28,18 @@ const HeaderForm = ({ handleSearchSubmit }) => {
   const [domainError, setDomainError] = useState(false);
   const [diplomaError, setDiplomaError] = useState(false);
 
+  const jobChanged = async function (val, setLoadingState) {
+    let res = await domainChanged(val, setDomainError)
+    setLoadingState('done')
+    return res;
+  };
+
+  const addressChanged = async function (val, setLoadingState) {
+    let res = await fetchAddresses(val)
+    setLoadingState('done')
+    return res
+  }
+
   const renderFormik = () => {
     return (
       <Formik
@@ -48,10 +60,11 @@ const HeaderForm = ({ handleSearchSubmit }) => {
                   setDiplomas
                 )}
                 compareItemFunction={compareAutoCompleteValues}
-                onInputValueChangeFunction={partialRight(domainChanged, setDomainError)}
+                onInputValueChangeFunction={jobChanged}
                 previouslySelectedItem={formValues?.job ?? null}
                 name="jobField"
                 placeholder="Ex : boulangerie"
+                searchPlaceholder="Indiquez le métier recherché ci-dessus"
               />
             </div>
             <div className="ml-3">
@@ -62,10 +75,11 @@ const HeaderForm = ({ handleSearchSubmit }) => {
                   itemToStringFunction={autoCompleteToStringFunction}
                   onSelectedItemChangeFunction={partialRight(formikUpdateValue, "location")}
                   compareItemFunction={compareAutoCompleteValues}
-                  onInputValueChangeFunction={fetchAddresses}
+                  onInputValueChangeFunction={addressChanged}
                   previouslySelectedItem={formValues?.location ?? null}
                   name="placeField"
                   placeholder="Adresse, ville ou code postal"
+                  searchPlaceholder="Indiquez le lieu recherché ci-dessus"
                 />
               </div>
             </div>