+++ ui/components/SearchForTrainingsAndJobs/components/SearchForm.js
@@ -25,6 +25,18 @@ const SearchForm = (props) => {
   const [domainError, setDomainError] = useState(false);
   const [diplomaError, setDiplomaError] = useState(false);
 
+  const jobChanged = async function (val, setLoadingState) {
+    let res = await domainChanged(val, setDomainError)
+    setLoadingState('done')
+    return res;
+  };
+
+  const addressChanged = async function (val, setLoadingState) {
+    let res = await fetchAddresses(val)
+    setLoadingState('done')
+    return res
+  }
+
   const renderFormik = () => {
     return (
       <Formik
@@ -59,10 +71,11 @@ const SearchForm = (props) => {
                             setDiplomas
                           )}
                           compareItemFunction={compareAutoCompleteValues}
-                          onInputValueChangeFunction={partialRight(domainChanged, setDomainError)}
+                          onInputValueChangeFunction={jobChanged}
                           previouslySelectedItem={formValues?.job ?? null}
                           name="jobField"
                           placeholder="Ex : boulangerie"
+                          searchPlaceholder="Indiquez le métier recherché ci-dessus"
                         />
                         <ErrorMessage name="job" className="errorField" component="div" />
                       </div>
@@ -79,11 +92,12 @@ const SearchForm = (props) => {
                     itemToStringFunction={autoCompleteToStringFunction}
                     onSelectedItemChangeFunction={partialRight(formikUpdateValue, "location")}
                     compareItemFunction={compareAutoCompleteValues}
-                    onInputValueChangeFunction={fetchAddresses}
+                    onInputValueChangeFunction={addressChanged}
                     previouslySelectedItem={formValues?.location ?? null}
                     scrollParentId="choiceColumn"
                     name="placeField"
                     placeholder="Adresse, ville ou code postal"
+                    searchPlaceholder="Indiquez le lieu recherché ci-dessus"
                   />
                   <ErrorMessage name="location" className="errorField" component="div" />
                 </div>