+++ server/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## 1.13.3 [05/05/2021]
 [610] Envoi event vers catalogue à l'affichage d'une formation
+[515] Page tampon accès recruteur
 
 ## 1.13.2 [05/05/2021]
 [609] Masquage sessions déjà commencées, meilleure présentation formations