+++ server/CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 1.14.0 [07/05/2021]
+[559] Statut autocomplete
+
 ## 1.13.4 [07/05/2021]
 [595] tracking gtm
 [579] Suppression notion de poids sur les opportunités d'emploi